+++ astropy/analytic_functions/__init__.py
@@ -1,3 +1,10 @@
 # Licensed under a 3-clause BSD style license - see LICENSE.rst
+"""This package contains analytic functions useful for astronomy.
 
-from .blackbody import *
+In future versions of ``astropy``, many of these might be
+accessible as `~astropy.modeling.core.Model`.
+
+"""
+
+# Shortcuts for most commonly used blackbody functions
+from .blackbody import blackbody_nu, blackbody_lambda