+++ astropy/analytic_functions/blackbody.py
@@ -3,32 +3,83 @@
 from __future__ import (absolute_import, division, print_function,
                         unicode_literals)
 
+# STDLIB
+import warnings
+
 # THIRD-PARTY
 import numpy as np
 
 # LOCAL
 from .. import constants as const
 from .. import units as u
+from ..utils.exceptions import AstropyUserWarning
 
 
-__all__ = ['planck']
+__all__ = ['blackbody_nu', 'blackbody_lambda']
 
-# Underflow and overflow limits
-_VERY_SMALL = 1e-4
-_VERY_LARGE = 85.0
-
-# Constants and units
-HC = const.h.cgs * const.c.to('AA/s')
+# Units
 FNU = u.erg / (u.cm**2 * u.s * u.Hz)
+FLAM = u.erg / (u.cm**2 * u.s * u.AA)
 
 
-def planck(in_x, temperature, flux_unit=FNU, silence_numpy=True):
-    """Calculate blackbody flux per steradian.
+def blackbody_nu(in_x, temperature):
+    """Calculate blackbody flux per steradian, :math:`B_{\\nu}(T)`.
+
+    .. note::
+
+        Use `numpy.errstate` to suppress Numpy warnings, if desired.
 
     .. warning::
 
-        Data points where overflow or underflow occurs will be set
-        to zeroes.
+        Output values might contain ``nan`` and ``inf``.
+
+    Parameters
+    ----------
+    in_x : number, array_like, or `~astropy.units.Quantity`
+        Frequency, wavelength, or wave number.
+        If not a Quantity, it is assumed to be in Hz.
+
+    temperature : number or `~astropy.units.Quantity`
+        Blackbody temperature.
+        If not a Quantity, it is assumed to be in Kelvin.
+
+    Returns
+    -------
+    flux : `~astropy.units.Quantity`
+        Blackbody monochromatic flux in
+        :math:`erg \\; cm^{-2} s^{-1} Hz^{-1} sr^{-1}`.
+
+    Raises
+    ------
+    ValueError
+        Invalid temperature.
+
+    ZeroDivisionError
+        Wavelength is zero (when converting to frequency).
+
+    """
+    # Convert to units for calculations, also force double precision
+    with u.add_enabled_equivalencies(u.spectral() + u.temperature()):
+        freq = u.Quantity(in_x, u.Hz, dtype=np.float64)
+        temp = u.Quantity(temperature, u.K, dtype=np.float64)
+
+    # Check if input values are physically possible
+    if temp < 0:
+        raise ValueError('Invalid temperature {0}'.format(temp))
+    if np.any(freq <= 0):  # pragma: no cover
+        warnings.warn('Input contains invalid wavelength/frequency value(s)',
+                      AstropyUserWarning)
+
+    # Calculate blackbody flux
+    bb_nu = (2.0 * const.h * freq ** 3 /
+             (const.c ** 2 * np.expm1(const.h * freq / (const.k_B * temp))))
+    flux = bb_nu.to(FNU, u.spectral_density(freq))
+
+    return flux / u.sr  # Add per steradian to output flux unit
+
+
+def blackbody_lambda(in_x, temperature):
+    """Like :func:`blackbody_nu` but for :math:`B_{\\lambda}(T)`.
 
     Parameters
     ----------
@@ -36,70 +87,21 @@ def planck(in_x, temperature, flux_unit=FNU, silence_numpy=True):
         Frequency, wavelength, or wave number.
         If not a Quantity, it is assumed to be in Angstrom.
 
-    temp : number or `~astropy.units.Quantity`
+    temperature : number or `~astropy.units.Quantity`
         Blackbody temperature.
         If not a Quantity, it is assumed to be in Kelvin.
-
-    flux_unit : `~astropy.units.Unit`
-        Flux unit for the blackbody radiation.
-        By default, it is :math:`erg \\; cm^{-2} s^{-1} Hz^{-1}`.
-        Any given unit must be convertible from the default with
-        :func:`~astropy.units.equivalencies.spectral_density`.
-
-    silence_numpy : bool
-        By default, this function temporarily silences Numpy warnings
-        about overflow and underflow because they are already handled
-        inside the function.
 
     Returns
     -------
     flux : `~astropy.units.Quantity`
-        Blackbody monochromatic flux in given unit per steradian.
+        Blackbody monochromatic flux in
+        :math:`erg \\; cm^{-2} s^{-1} \\AA^{-1} sr^{-1}`.
 
     """
-    # Temporarily silence Numpy
-    if silence_numpy:
-        old_np_err_cfg = np.seterr(all='ignore')
-
-    # Convert to Angstrom for calculations
     if not isinstance(in_x, u.Quantity):
-        w = in_x * u.AA
-    else:
-        w = in_x.to(u.AA, u.spectral())
+        in_x = in_x * u.AA
 
-    # Convert to Kelvin for calculations
-    if not isinstance(temperature, u.Quantity):
-        temp = temperature * u.K
-    else:
-        temp = temperature.to(u.K, u.temperature())
+    bb_nu = blackbody_nu(in_x, temperature) * u.sr  # Remove sr for conversion
+    flux = bb_nu.to(FLAM, u.spectral_density(in_x))
 
-    # Force double precision
-    w = w.astype(np.float64)
-    temp = temp.astype(np.float64)
-
-    x = w * temp
-
-    # Catch division by zero
-    mask = x > 0
-    x = np.where(mask, HC / (const.k_B.cgs * x), 0.0)
-
-    # Catch overflow/underflow
-    mask = (x >= _VERY_SMALL) & (x < _VERY_LARGE)
-    factor = np.where(mask, 1.0 / np.expm1(x), 0.0)
-
-    # Calculate blackbody flux
-    freq = w.to(u.Hz, u.spectral())
-    flux = 2.0 * const.h * factor * freq * freq * freq / const.c ** 2
-    flux = flux.to(flux_unit, u.spectral_density(w))
-
-    # Catch NaN
-    flux[factor == 0] = 0.0
-
-    # Add per steradian to output flux unit
-    flux /= 1 * u.sr
-
-    # Restore Numpy settings
-    if silence_numpy:
-        np.seterr(**old_np_err_cfg)
-
-    return flux
+    return flux / u.sr  # Add per steradian to output flux unit